Moves the getStepWidth function to the inputs module
