+++ src/inputs.py
@@ -29,3 +29,20 @@ def getNumberOfLinesFromUser():
                 print("ERROR - Number of lines must be in the range [3-80].")
         
     return numberOfLines
+
+def getStepWidth():
+    rawStepWidth = ""
+    stepWidth = 0
+
+    while len(rawStepWidth) == 0 or stepWidth < 1 or stepWidth > 10 or not numberIsAWholeNumber(rawStepWidth):
+        rawStepWidth = input("Width of each step? ")
+        if len(rawStepWidth) == 0:
+            print("ERROR - Input must not be empty.")
+        elif not numberIsAWholeNumber(rawStepWidth):
+            print("ERROR - Input must be a whole positive number.")
+        else:
+            stepWidth = int(rawStepWidth)
+            if stepWidth < 1 or stepWidth > 10:
+                print("ERROR - Step width must be in the range [1-10].")
+
+    return stepWidth
